requirements.txt
streamlit==1.38.0
Pillow==10.4.0
pytesseract==0.3.13
deep-translator==1.11.4
langdetect==1.0.9